nextapp/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
polygon-api-client==1.12.0
google-generativeai==0.3.2
flask==3.0.2
flask-cors==4.0.0 